+++ client/shared/IPC/include/Client/Shared/IPC/Message/Logon.hpp
@@ -8,7 +8,7 @@ namespace ipc::msg
     {
         std::string UserId;
         uint32_t SeqNum;
-        float HeartBeat;
+        uint32_t HeartBeat;
 
         friend net::Buffer &operator<<(net::Buffer &_buffer, const Logon &_logon);
         friend net::Buffer &operator>>(net::Buffer &_buffer, Logon &_logon);
